CPM-1016: Move into Property folder
