+++ components/identifier-generator/back/tests/Acceptance/Context/CreateIdentifierGeneratorContext.php
@@ -289,79 +289,6 @@ final class CreateIdentifierGeneratorContext implements Context
         $this->tryToCreateGenerator(conditions: [$defaultCondition]);
     }
 
-    /**
-     * @When /^I try to create an identifier generator with a simple select property without attribute code$/
-     */
-    public function iTryToCreateAnIdentifierGeneratorWithASimpleSelectPropertyWithoutAttributeCode(): void
-    {
-        $this->tryToCreateGenerator(structure: [
-            ['type' => 'simple_select', 'process' => ['type' => 'no']],
-        ]);
-    }
-
-    /**
-     * @When /^I try to create an identifier generator with simple select property without process field$/
-     */
-    public function iTryToCreateAnIdentifierGeneratorWithSimpleSelectPropertyWithoutProcessField(): void
-    {
-        $this->tryToCreateGenerator(structure: [
-            ['type' => 'simple_select', 'attributeCode' => 'color'],
-        ]);
-    }
-
-    /**
-     * @When /^I try to create an identifier generator with a simple_select property with (?P<attributeCode>[^']*) attribute(?: and (?P<scope>.*) scope)?(?: and (?P<locale>.*) locale)?$/
-     */
-    public function iTryToCreateAnIdentifierGeneratorWithASimpleSelectPropertyWithNameAttribute(
-        string $attributeCode,
-        string $scope = '',
-        string $locale = ''
-    ): void {
-        $simpleSelectProperty = ['type' => 'simple_select', 'attributeCode' => $attributeCode, 'process' => ['type' => 'no']];
-
-        if ($scope) {
-            $simpleSelectProperty['scope'] = $scope;
-        }
-
-        if ($locale) {
-            $simpleSelectProperty['locale'] = $locale;
-        }
-
-        $this->tryToCreateGenerator(structure: [
-            $simpleSelectProperty,
-        ]);
-    }
-
-    /**
-     * @When /^I try to create an identifier generator with a simple select process with type (?P<type>[^']*) and operator (?P<operator>[^']*) and (?P<value>[^']*) as value$/
-     */
-    public function iTryToCreateAnIdentifierGeneratorWithSimpleSelectProcessWithTypeAndOperatorAndValue($type, $operator, $value): void
-    {
-        $value = \json_decode($value);
-        $defaultStructure = [
-            'attributeCode' => 'color',
-            'type' => 'simple_select',
-            'process' => ['type' => $type, 'operator' => $operator, 'value' => $value],
-        ];
-        if ($operator === 'undefined') {
-            unset($defaultStructure['process']['operator']);
-        }
-        if ($value === 'undefined') {
-            unset($defaultStructure['process']['value']);
-        }
-        $this->tryToCreateGenerator(structure: [$defaultStructure]);
-    }
-
-    /**
-     * @Then there should be no :localeCode label for the :generatorCode generator
-     */
-    public function thereShouldBeNoLabelForLocale(string $localeCode, string $generatorCode): void
-    {
-        $identifierGenerator = $this->generatorRepository->get($generatorCode);
-        Assert::isInstanceOf($identifierGenerator, IdentifierGenerator::class);
-        Assert::keyNotExists($identifierGenerator->labelCollection()->normalize(), $localeCode);
-    }
-
     private function tryToCreateGenerator(
         ?string $code = null,
         ?array $structure = null,
